// src/serverApi/v3/models/task-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { RichText } from './rich-text';
import { TaskStatusResponse } from './task-status-response';
import { UsersList } from './users-list';

/**
 * 
 * @export
 * @interface TaskResponse
 */
export interface TaskResponse {
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    id: string;
    /**
     * 
     * @type {Array<UsersList>}
     * @memberof TaskResponse
     */
    users: Array<UsersList>;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    availableDate?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    duedate?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    courseName: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    lessonName?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    courseId: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    taskCardId?: string;
    /**
     * Task description object, with props content: string and type: input format types
     * @type {RichText}
     * @memberof TaskResponse
     */
    description?: RichText;
    /**
     * 
     * @type {boolean}
     * @memberof TaskResponse
     */
    lessonHidden: boolean;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    displayColor?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    createdAt: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    updatedAt: string;
    /**
     * 
     * @type {TaskStatusResponse}
     * @memberof TaskResponse
     */
    status: TaskStatusResponse;
}